fix comments

+++ x/permission/keeper/keeper.go
@@ -360,28 +360,39 @@ func (k Keeper) DeletePolicy(ctx sdk.Context, principal *types.Principal, resour
 	return policyID, nil
 }
 
-// ForceDeleteAccountPolicyForResource deletes all individual accounts policy enforced on resources
-func (k Keeper) ForceDeleteAccountPolicyForResource(ctx sdk.Context, resourceType resource.ResourceType, resourceID math.Uint) {
+// ForceDeleteAccountPolicyForResource deletes all individual accounts policy enforced on resources, if
+func (k Keeper) ForceDeleteAccountPolicyForResource(ctx sdk.Context, maxDelete, deletedCount uint64, resourceType resource.ResourceType, resourceID math.Uint) (uint64, bool) {
 	if resourceType == resource.RESOURCE_TYPE_UNSPECIFIED {
-		return
+		return deletedCount, true
 	}
 	store := ctx.KVStore(k.storeKey)
 	resourceAccountsPolicyStore := prefix.NewStore(store, types.PolicyForAccountPrefix(resourceID, resourceType))
 	iterator := resourceAccountsPolicyStore.Iterator(nil, nil)
 	defer iterator.Close()
+
 	for ; iterator.Valid(); iterator.Next() {
+		// if exceeding the limit, pause the GC and mark the current resource's deletion is not complete yet
+		if deletedCount > maxDelete {
+			return deletedCount, false
+		}
 		// delete mapping policyId -> policy
 		policyID := sequence.DecodeSequence(iterator.Value())
 		store.Delete(types.GetPolicyByIDKey(policyID))
 		// delete mapping policyKey -> policyId
 		resourceAccountsPolicyStore.Delete(iterator.Key())
+		deletedCount++
+		// emit DeletePolicy Event
+		_ = ctx.EventManager().EmitTypedEvents(&types.EventDeletePolicy{
+			PolicyId: policyID,
+		})
 	}
+	return deletedCount, true
 }
 
 // ForceDeleteGroupPolicyForResource deletes group policy enforced on resource
-func (k Keeper) ForceDeleteGroupPolicyForResource(ctx sdk.Context, resourceType resource.ResourceType, resourceID math.Uint) {
+func (k Keeper) ForceDeleteGroupPolicyForResource(ctx sdk.Context, maxDelete, deletedTotal uint64, resourceType resource.ResourceType, resourceID math.Uint) (uint64, bool) {
 	if resourceType == resource.RESOURCE_TYPE_UNSPECIFIED || resourceType == resource.RESOURCE_TYPE_GROUP {
-		return
+		return deletedTotal, true
 	}
 	policyForGroupKey := types.GetPolicyForGroupKey(resourceID, resourceType)
 	store := ctx.KVStore(k.storeKey)
@@ -390,14 +401,23 @@ func (k Keeper) ForceDeleteGroupPolicyForResource(ctx sdk.Context, resourceType 
 		policyGroup := types.PolicyGroup{}
 		k.cdc.MustUnmarshal(bz, &policyGroup)
 		for i := 0; i < len(policyGroup.Items); i++ {
+			if deletedTotal > maxDelete {
+				return deletedTotal, false
+			}
 			// delete concrete policy by policyId
-			store.Delete(types.GetPolicyByIDKey(policyGroup.Items[i].PolicyId))
+			policyId := policyGroup.Items[i].PolicyId
+			store.Delete(types.GetPolicyByIDKey(policyId))
+			deletedTotal++
+			_ = ctx.EventManager().EmitTypedEvents(&types.EventDeletePolicy{
+				PolicyId: policyId,
+			})
 		}
 		store.Delete(policyForGroupKey)
 	}
+	return deletedTotal, true
 }
 
-// ForceDeleteGroupMembers deletes group members when user deletes group
+// ForceDeleteGroupMembers deletes group members when user deletes a group
 func (k Keeper) ForceDeleteGroupMembers(ctx sdk.Context, groupId math.Uint) {
 	store := ctx.KVStore(k.storeKey)
 	groupMembersPrefixStore := prefix.NewStore(ctx.KVStore(k.storeKey), types.GroupMembersPrefix(groupId))
@@ -410,4 +430,31 @@ func (k Keeper) ForceDeleteGroupMembers(ctx sdk.Context, groupId math.Uint) {
 		// delete GroupMemberPrefix_groupId_memberAddr -> memberSequence(id)
 		groupMembersPrefixStore.Delete(iter.Key())
 	}
+}
+
+func (k Keeper) ExistAccountPolicyForResource(ctx sdk.Context, resourceType resource.ResourceType, resourceID math.Uint) bool {
+	if resourceType == resource.RESOURCE_TYPE_UNSPECIFIED {
+		return false
+	}
+	store := ctx.KVStore(k.storeKey)
+	resourceAccountsPolicyStore := prefix.NewStore(store, types.PolicyForAccountPrefix(resourceID, resourceType))
+	iterator := resourceAccountsPolicyStore.Iterator(nil, nil)
+	defer iterator.Close()
+	return iterator.Valid()
+}
+
+func (k Keeper) ExistGroupPolicyForResource(ctx sdk.Context, resourceType resource.ResourceType, resourceID math.Uint) bool {
+	if resourceType == resource.RESOURCE_TYPE_UNSPECIFIED || resourceType == resource.RESOURCE_TYPE_GROUP {
+		return false
+	}
+	policyForGroupKey := types.GetPolicyForGroupKey(resourceID, resourceType)
+	store := ctx.KVStore(k.storeKey)
+	return store.Has(policyForGroupKey) && store.Get(policyForGroupKey) != nil
+}
+
+func (k Keeper) ExistGroupMemberForGroup(ctx sdk.Context, groupId math.Uint) bool {
+	groupMembersPrefixStore := prefix.NewStore(ctx.KVStore(k.storeKey), types.GroupMembersPrefix(groupId))
+	iter := groupMembersPrefixStore.Iterator(nil, nil)
+	defer iter.Close()
+	return iter.Valid()
 }
